test: refactor test structure
